changes in Order service

+++ Payment/src/main/java/com/payment/entity/Payment.java
@@ -22,7 +22,7 @@ public class Payment {
 	private String paymentStatus;
 	private String transactionId;
 	
-	private long amount;
+	private double amount;
 	private long orderId;
 	
 	@Override
@@ -48,10 +48,11 @@ public class Payment {
 	public void setTransactionId(String transactionId) {
 		this.transactionId = transactionId;
 	}
-	public long getAmount() {
+	
+	public double getAmount() {
 		return amount;
 	}
-	public void setAmount(long amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
 	public long getOrderId() {
